august/site/app/requirements.txt
pyJWT
flask
python-dotenv
Flask-Limiter
requests